updated profile and made ui cleaner
